Refactor all functions to put them in their respective files, refactor tests, add github actions for test, and add test badge
