+++ README.md
@@ -1,0 +1,60 @@
+# testing-practice
+![Linters](https://github.com/Bluette1/testing-practice/workflows/Linters/badge.svg?branch=feature%2Fadd-testing-practice)  ![Test with jest](https://github.com/Bluette1/testing-practice/workflows/Test%20with%20jest/badge.svg)
+
+## Project description
+This project was done mainly for learning purposes, the aim was to practice testing using jest, we created functions (arry-analysis, calculator, caesar-cipher, string-reversal, capitalize) and we created tests for them. The project's specifications are listed on [The odin project](https://www.theodinproject.com/courses/javascript/lessons/testing-practice)
+
+## Built with
+- Vanilla javascript
+- [Jest](https://jestjs.io)
+
+## Running and testing it locally
+
+### Prerequisites
+- [NodeJS](https://nodejs.org/)
+- [Yarn](https://yarnpkg.com/) or just the default npm which comes with NodeJS
+
+### Setup
+- Run `git clone https://github.com/Bluette1/testing-practice && cd testing-practice` to get a copy of source codes on your local computer and to navigate inside the project directory
+- Run `yarn` to install required dependencies
+
+### Tests
+- Run `yarn test` to run all tests suites
+
+## Contributions
+
+There are two ways of contributing to this project:
+
+1.  If you see something wrong or not working, please check [the issue tracker section](https://github.com/Bluette1/testing-practice/issues), if that problem you met is not in already opened issues then open a new issue by clicking on `new issue` button.
+
+2.  If you have a solution to that, and you are willing to work on it, follow the below steps to contribute:
+    1.  Fork this repository
+    1.  Clone it on your local computer by running `git clone https://github.com/your-username/testing-practice.git` __Replace *your username* with the username you use on github__
+    1.  Open the cloned repository which appears as a folder on your local computer with your favorite code editor
+    1.  Create a separate branch off the *master branch*,
+    1.  Write your codes which fix the issue you found
+    1.  Commit and push the branch you created
+    1.  Raise a pull request, comparing your new created branch with our original master branch [here](https://github.com/Bluette1/testing-practice/testing-practice)
+
+## Author
+
+👤 **Mugirase Emmanuel**
+
+- Github: [@descholar-ceo](https://github.com/descholar-ceo)
+- Twitter: [@descholar3](https://twitter.com/descholar3)
+- Linkedin: [MUGIRASE Emmanuel](https://www.linkedin.com/in/mugirase-emmanuel)
+
+👤 **Marylene Sawyer**
+- Github: [@Bluette1](https://github.com/Bluette1)
+- Twitter: [@MaryleneSawyer](https://twitter.com/MaryleneSawyer)
+- Linkedin: [Marylene Sawyer](https://www.linkedin.com/in/marylene-sawyer)
+
+
+## Show your support
+
+Give a ⭐️ if you like this project!
+
+## Acknowledgment
+- [Microverse](https://www.microvese.org)
+- [Webpack](https://webpack.js.org/)
+
