+++ src/calculator.js
@@ -1,0 +1,11 @@
+const calculator = {
+  add: (first, sec) => first + sec,
+
+  subtract: (first, sec) => first - sec,
+
+  divide: (first, sec) => first / sec,
+
+  multiply: (first, sec) => first * sec,
+};
+
+export default calculator;
